The classes VoltageDividerR1 and VoltageDividerR2 are now in a single file VoltageDivider.cpp The definitions for bot classes are now in VoltageDivider.h
